cleaned up comments

+++ code/include/Storage.h
@@ -24,6 +24,10 @@
 // ------------------------ RunManager ---------------------
 // =========================================================
 
+
+/**
+ * @brief RunManager class to manage runs in a storage device
+ */
 class RunManager {
   private:
     std::string baseDir;
@@ -31,23 +35,43 @@ class RunManager {
     int nextRunIndex = 0; // next run id
     RowCount totalRecords = 0;
 
+    /**
+     * @brief reads the run files from the baseDir and returns the list of run files.
+     */
     std::vector<std::string> getRunInfoFromDir();
 
   public:
+    /**
+     * @brief RunManager class to manage runs in a storage device.
+     * It creates a directory for the device runs.
+     * If the directory exists, it deletes all run files in the directory.
+     * @param deviceName
+     */
     RunManager(std::string deviceName);
+
+    /**
+     * @brief warns the user if all run files are not deleted
+     */
     ~RunManager();
 
-    // setters
+    // Setters
     void addRunFile(std::string filename, RowCount nRecords) {
         runFiles.push_back({filename, nRecords});
         totalRecords += nRecords;
     }
 
+    /**
+     * @brief removes the run file from the runFiles list.
+     * It also updates the totalRecords count.
+     */
     bool removeRunFile(std::string filename) {
+        // Find the run file
         auto it = std::find_if(
             runFiles.begin(), runFiles.end(),
             [filename](const std::pair<std::string, RowCount> &p) { return p.first == filename; });
+
         if (it != runFiles.end()) {
+            // Remove the run file and update the total records count
             totalRecords -= it->second;
             runFiles.erase(it);
             return true;
@@ -55,7 +79,7 @@ class RunManager {
         return false;
     }
 
-    // getters
+    // Getters
     std::string getBaseDir() { return baseDir; }
     RowCount getTotalRecords() { return totalRecords; }
     std::string getNextRunFileName();
@@ -107,14 +131,30 @@ class Storage {
     Storage *spillTo = nullptr; // used for merging
     RunWriter *spillWriter = nullptr;
 
-    // setup
+    /**
+     * @brief Constructor for the Storage class
+     */
     Storage(std::string name, ByteCount capacity, int bandwidth, double latency);
+
+    /**
+     * @brief Configure the storage device based on the provided configurations.
+     * - Calculate the page size in records
+     * - Calculate the merge fan-in and merge fan-out in records and the cluster size
+     * - Create a run manager for HDD and SSD
+     */
     void configure();
+
+    /**
+     * @brief Set the spillTo storage device for spilling when this storage device is full.
+     */
     void setSpillTo(Storage *spillTo) {
         this->spillTo = spillTo;
         printv("%s set SpillTo to %s\n", this->getName().c_str(), spillTo->getName().c_str());
     }
 
+    /**
+     * @brief Destructor for the Storage class
+     */
     ~Storage() {
         if (runManager != nullptr) {
             delete runManager;
@@ -138,9 +178,10 @@ class Storage {
     double getAccessTimeInSec(RowCount nRecords) const {
         return this->LATENCY + (nRecords * Config::RECORD_SIZE) / this->BANDWIDTH;
     }
-    int getAccessTimeInMillis(RowCount nRecords) const {
-        return (int)(this->getAccessTimeInSec(nRecords) * 1000);
+    double getAccessTimeInMillis(RowCount nRecords) const {
+        return (this->getAccessTimeInSec(nRecords) * 1000);
     }
+
     // ----------------------------- space calculations ------------------------
     // getters
     RowCount getTotalEmptySpaceInRecords() {
@@ -216,11 +257,6 @@ class Storage {
         _totalSpaceInInputClusters = 0;
         _totalSpaceInOutputClusters = 0;
     }
-    // virtual void setupMergeState(RowCount outputDevicePageSize, int fanIn) = 0;
-    // /**
-    //  * @brief Setup merging state for dram (used for merging miniruns)
-    //  */
-    // virtual int setupMergeStateForMiniruns(RowCount outputDevicePageSize) = 0;
 
     // --------------------------- FILE I/O ------------------------------------
     std::string getReadFilePath() const { return readFilePath; }
